==> 1_get_project_tracker_data.ipynb <==
import arcpy
from arcpy import env
import os
import arcgis
from arcgis.gis import GIS
from MTIPmapping import *

path = r'T:\MPO\TIP\TIP FY24-27\Maps\Data'
arcpy.env.workspace = os.path.join(path, 'Data.gdb')
arcpy.env.overwriteOutput = True

write_json('https://gis.odot.state.or.us/arcgis1006/rest/services/tpt/tpt_display/MapServer/1?f=pjson',
        'all_lines')

# this doesn't include spatial data
#json2shp('all_lines')

gis = GIS('home')

gis.users.me.username

items = gis.content.search(query="owner:" + gis.users.me.username + " AND title:Project Tracker",
                               item_type="Web Map")

items
