==> 2_review_project_data.ipynb <==
import pandas as pd
import geopandas as gpd
import os

path = r'T:\MPO\TIP\TIP FY24-27\Maps\Data'

file = path + '\\MTIP24_27_List.xlsx'

data = pd.read_excel(file)

data.head(3)

data.columns

data.shape

len(data['MTIP ID #'].unique())

len(data['STIP Key'].unique())

lines = gpd.read_file(path+'\\TPT_for_lcog.gdb', layer='all_lines')

lines.columns

points = gpd.read_file(path+'\\TPT_for_lcog.gdb', layer='all_points')

points.columns

data['MTIP ID #'].dropna().values

stip_keys = list(data['STIP Key'].dropna().unique())

pkeys = list(map(int, points['FME_PROJ_KEY_NO'].values))

type(pkeys[0])

type(stip_keys[0])

lkeys = list(map(int, lines['FME_PROJ_KEY_NO'].values))

loc_keys_p = list(points['STIP_LOC_ID'].values)

loc_keys_l = list(lines['STIP_LOC_ID'].values)

lid = [key for key in stip_keys if key in lkeys]

pid = [key for key in stip_keys if key in pkeys]

len(pid)

print(list(map(str, lid)))

print(list(map(str, pid)))

selected_points = points[points.FME_PROJ_KEY_NO.isin(list(map(str, pid)))][['FME_PROJ_KEY_NO', 'STIP_LOC_ID', 'geometry']]

selected_points.rename(columns={'FME_PROJ_KEY_NO':'STIP_Key', 'STIP_LOC_ID':'LOC_ID'}, inplace=True)

selected_lines = lines[lines.FME_PROJ_KEY_NO.isin(list(map(str, lid)))][['FME_PROJ_KEY_NO', 'STIP_LOC_ID','geometry']]

selected_lines.rename(columns={'FME_PROJ_KEY_NO':'STIP_Key', 'STIP_LOC_ID':'LOC_ID'}, inplace=True)

selected_points.columns

selected_points.to_file(os.path.join(path, 'ODOT_points.shp'))

selected_lines.to_file(os.path.join(path, 'ODOT_lines.shp'))

mapped_ids = lid + pid

mapped = data[data['STIP Key'].isin(mapped_ids)]

mapped.to_csv(os.path.join(path, 'mapped_proj.csv'), index=False)

unmapped = data[~(data['STIP Key'].isin(mapped_ids))]

unmapped.to_csv(os.path.join(path, 'unmapped_proj.csv'), index=False)
