==> 3_combine_all_mapped_data.ipynb <==
import geopandas as gpd
import pandas as pd
import os
import numpy as np

path = r'T:\MPO\TIP\TIP FY24-27\Maps\Data'

mapped = pd.read_csv(os.path.join(path, 'mapped_proj.csv'))

unmapped = pd.read_csv(os.path.join(path, 'unmapped_proj.csv'))

mapped.shape

mapped.columns

unmapped.columns

FY24_27Lines = gpd.read_file(os.path.join(path, 'Data.gdb'), layer='FY24_27Lines')

FY24_27Lines.head(3)

ODOTlines.head(3)

FY24_27Points = gpd.read_file(os.path.join(path, 'Data.gdb'), layer='FY24_27Points')

lines = gpd.read_file(path+'\\TPT_for_lcog.gdb', layer='all_lines')

lines.columns

selected_line = lines[lines.FME_PROJ_KEY_NO == '19917']

selected_line = selected_line[['FME_PROJ_KEY_NO', 'STIP_LOC_ID', 'geometry']]

selected_line.rename(columns={'FME_PROJ_KEY_NO':'STIP_Key', 'STIP_LOC_ID':'LOC_ID'}, inplace=True)

selected_line['STIP_Key'] = '22696'

selected_line

ODOTlines = gpd.read_file(os.path.join(path, 'ODOT_lines.shp'))

ODOTlines = pd.concat([ODOTlines, selected_line], ignore_index=True)

ODOTpoints = gpd.read_file(os.path.join(path, 'ODOT_points.shp'))

mapped_keys = mapped[~mapped.Link.isnull()].STIP_Key.unique()

unmapped_keys = unmapped[~unmapped['Mapped?'].isnull()].STIP_Key.unique()

all_mapped_keys = np.concatenate((mapped_keys, unmapped_keys))

LineKeys = FY24_27Lines.STIP_Key.unique()

PointKeys = FY24_27Points.STIP_Key.unique()

keys = np.concatenate((LineKeys, PointKeys))

ODOTlines.head(3)

ODOTkeys = np.concatenate((ODOTlines.STIP_Key.unique(), ODOTpoints.STIP_Key.unique()))

ODOTkeys.astype(np.int64)

[key for key in keys if key not in all_mapped_keys]

# correct this on the table
22723 in unmapped.STIP_Key.unique()

keys_to_add = list(map(str, [key for key in ODOTkeys.astype(np.int64) if key not in keys]))

MTIPlines = pd.concat([FY24_27Lines[['STIP_Key','geometry']], 
           ODOTlines[ODOTlines.STIP_Key.isin(keys_to_add)][['STIP_Key','geometry']]], 
          ignore_index=True)

MTIPpoints = pd.concat([FY24_27Points[['STIP_Key','geometry']], 
           ODOTpoints[ODOTpoints.STIP_Key.isin(keys_to_add)][['STIP_Key','geometry']]], 
          ignore_index=True)

ODOTlines[ODOTlines.STIP_Key.isin(keys_to_add)][['STIP_Key','geometry']]

MTIPlines[MTIPlines.STIP_Key != 23058].to_file(os.path.join(path, 'FY24_27_lines.shp'))

MTIPlines.STIP_Key.dtype

MTIPpoints.to_file(os.path.join(path, 'FY24_27_points.shp'))
