==> 6_add_more_spatial_data.ipynb <==
import pandas as pd
import geopandas as gpd
import numpy as np
from geopandas import GeoDataFrame
from shapely.geometry import Point
from geopy.geocoders import ArcGIS
import os

path = r'T:\MPO\TIP\TIP FY24-27\Maps\Data'

df = pd.read_excel(os.path.join(path, 'AddCoordinates.xlsx'))

df

df['ID'] = list(range(32, 32 + df.shape[0]))

geometry = [Point(xy) for xy in zip(df.Longitude, df.Latitude)]
gdf = GeoDataFrame(df, crs="EPSG:4326", geometry=geometry)

points = gpd.read_file(os.path.join(path, 'FY24_27_points.shp'))

gdf.rename(columns={'STIP Key':'STIP_Key'}, inplace=True)

gdf = gdf.to_crs(epsg=2992)

points_added = points.append(gdf[points.columns], ignore_index=True)

locs = ["Pioneer Pkwy E & E St, Springfield, OR 97477", 
        "E St & Pioneer Pkwy W, Springfield, OR 97477", 
        "69th St & Thurston Rd, Springfield, OR 97478"]

def getCoordinates(address):
    lonlat = ArcGIS().geocode(address)
    #latitude = lonlat[1][0]
    #longitude = lonlat[1][1]
    (latitude, longitude) = lonlat[1]
    return [address, longitude, latitude] 

# Create the pandas DataFrame
ndf = pd.DataFrame(dt, columns = ['Address', 'Longitude', 'Latitude'])

ndf['STIP_Key'] = 22351
ndf['ID'] = 16

geometry = [Point(xy) for xy in zip(ndf.Longitude, ndf.Latitude)]
ngdf = GeoDataFrame(ndf, crs="EPSG:4326", geometry=geometry)

ngdf = ngdf.to_crs(epsg=2992)

npoints_added = points_added.append(ngdf[points.columns], ignore_index=True)

data = pd.read_excel(os.path.join(path, 'MTIP24_27_List_Updated.xlsx'))

lines = gpd.read_file(os.path.join(path, 'FY24_27_lines.shp'))

Ids = npoints_added[['STIP_Key', 'ID']].drop_duplicates().append(lines[['STIP_Key', 'ID']].drop_duplicates(), ignore_index=True)

Ids.rename(columns={'STIP_Key':'STIP Key'}, inplace=True)

Ids['STIP Key'] = Ids['STIP Key'].astype(np.int64)

data.shape

ndata = data.merge(Ids, how='outer', on='STIP Key')

ndata.shape

ndata.to_excel(os.path.join(path, 'MTIP24_27_List_IDs.xlsx'), 
               sheet_name='2024-2027 Project List', 
               index=False)  

# overwrite the existing data
npoints_added.to_file(os.path.join(path, 'FY24_27_points.shp'))
