==> 8_reorganize_project_IDs_and_table.ipynb <==
import pandas as pd
import numpy as np
import geopandas as gpd
import os
from sqlalchemy import create_engine
import re
from MTIPmapping import *

path = r'T:\MPO\TIP\TIP FY24-27\Maps\Data'
points = gpd.read_file(os.path.join(path, 'FY24_27_points.shp'))
lines = gpd.read_file(os.path.join(path, 'FY24_27_lines.shp'))

[k for k in points.STIP_Key.unique() if k in lines.STIP_Key.unique()]

points.loc[points.STIP_Key=='22348', 'ID'].unique()

data = pd.read_excel(r'T:\MPO\TIP\TIP FY24-27\Maps\MTIP_FY24-27_20230303.xlsx')

data.shape

Ids = points[['STIP_Key', 'ID']].drop_duplicates().append(lines[['STIP_Key', 'ID']].drop_duplicates(), ignore_index=True)
Ids.rename(columns={'STIP_Key':'STIP Key'}, inplace=True)

data.columns

Ids['STIP Key'] = Ids['STIP Key'].astype(np.int64)

ndata = data.merge(Ids, how='outer', on='STIP Key')

23058 in ndata['STIP Key'].unique()

sdata = ndata[ndata['STIP Key'] != 23058]

sdata.loc[:,'AQ Exempt?'] = sdata.loc[:,'Air Quality Status'].apply(lambda x: getAQvars(x, var='AQ Exempt?')) 
sdata.loc[:,'AQ Status'] = sdata.loc[:,'Air Quality Status'].apply(lambda x: getAQvars(x, var='AQ Status'))
sdata.loc[:,'IAC'] = sdata.loc[:,'Air Quality Status'].apply(lambda x: getAQvars(x, var='IAC')) 

pointdata = get_district_IDs(dat=points, shptype='point', keycol='STIP_Key', 
                             colnm1='Senator District', colnm2='Representative District')
linedata = get_district_IDs(dat=lines, shptype='line', keycol='STIP_Key', 
                             colnm1='Senator District', colnm2='Representative District')

df = pointdata.append(linedata , ignore_index=True)
df.rename(columns={'STIP_Key':'STIP Key'}, inplace=True)
df['STIP Key'] = df['STIP Key'].astype(np.int64)

ndata = sdata.merge(df, how='outer', on='STIP Key')

ndata.shape

ndata.to_excel(os.path.join(path, 'MTIP24_27_List_20230303.xlsx'), 
               sheet_name='2024-2027 Project List', 
               index=False)
